chore: add ability to test entry paths
